+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to **Pipecat Cloud** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.16] - 2025-12-18
+
+### Fixed
+
+- Fixed an issue where deployments would appear to be ready right away. Now,
+  the deployment readiness matches that of the service.
+
 ## [0.2.15] - 2025-12-09
 
 ### Changed